+++ CHANGELOG.md
@@ -170,6 +170,8 @@ Full list of changes:
   Use `persistRequestsKey`.
 - Removed the `The function passed to Apify.main() threw an exception` error message,
   because it was confusing to users.
+- Removed automatic injection of `charset=utf-8` in `keyValueStore.setValue()`
+  to the `contentType` option.
 
 0.22.0 / 2020/12/07
 ====================